Game I did during my Micro-controller & Assembly Curse
In this code you will find a game I wrote in assembly of the 8086 processor.

In the game, you are an "O" symbol on the screen.
Your job is to "eat" the "X" symbol.
Every time you ean an X, your score total will increase by one.

Note that to run the code in your local mechine, you need a DOSBOX simulator of the 8086 proccesor.